moolmool/src/test/java/com/wappl/moolmool/qna/vo/AnswerVOTest.java
package com.wappl.moolmool.qna.vo;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;

import java.util.Set;

import javax.validation.ConstraintViolation;
import javax.validation.Validation;
import javax.validation.Validator;
import javax.validation.ValidatorFactory;

import org.junit.BeforeClass;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.transaction.annotation.Transactional;

@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(locations = { "file:src/main/webapp/WEB-INF/spring/root-context.xml" })
@Transactional
public class AnswerVOTest {
	private static Validator validator;
	
	@BeforeClass
	public static void setUpValidator() {
		ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
		validator = factory.getValidator();
	}
	
	@Test
	public void contentIsBlank() throws Exception {
		AnswerVO vo = new AnswerVO();
		vo.setContent("");
		Set<ConstraintViolation<AnswerVO>> constraintViolation =
				validator.validate(vo);
		assertThat(constraintViolation.size(), is(1));
	}
}
